Nuke onnx models from test  (#6890)
* Elimitated usage of onnx models in tests/ disabled some tests

* Verified green: batch 1

* Verified tests: batch 2

* Verified tests: batch 3

* Verified tests: batch 4

* Verified tests: batch 5

* Verified tests: batch 6

* Verified tests: batch 7

* Verified tests: batch 8

* Verified tests: batch 9

* Verified tests: batch 10

* Verified tests: batch 11

* Verified tests: batch 12

* Verified tests: batch 12_follow up

* Verified tests: batch 13

* Verified tests: batch 14

* Verified tests: batch 15

* Verified tests: batch 16

* Removed exits

+++ qa/L0_compute_capability/test.sh
@@ -53,6 +53,9 @@ RET=0
 BACKENDS=${BACKENDS:="graphdef savedmodel onnx libtorch plan"}
 
 for BACKEND in $BACKENDS; do
+    if [[ "$BACKEND" == 'onnx' ]]; then
+        continue
+    fi
     # Need just one model for the backend...
     rm -fr models && mkdir models
     cp -r ${DATADIR}/qa_model_repository/${BACKEND}_float32_float32_float32 \
